Add files via upload
Runs after bokeh-usage.ps1.  Takes the amount of lines in the usage file and subtracts it by the hard coded number of licenses owned to create a file saying how many licenses are available.  At the end it checks a file to see if the number of licenses in use is higher than what is in the file.  If it is higher it overwrites the file with the new number.