Add files via upload
Runs the licensing program for PG Bokeh with the license as an argument to get a list of current users who have taken a license. It creates a text file with the users and sends the users to a Database.  A web server on the same server as the Database displays the licensing information as well as who is using it.